Add missing LSS header
As detailed
[here](https://chromium.googlesource.com/breakpad/breakpad/+/master/docs/linux_starter_guide.md),
in order to build Breakpad on Linux the LSS header file is needed. This
file is supposed to be included as a submodule, but for convenience it
is instead added explicitly to this repository.

This file is taken from [the official LSS
repository](https://chromium.googlesource.com/linux-syscall-support/) at
commit `a89bf7903f3169e6bc7b8efc10a73a7571de21cf`.